+++ CHANGELOG.md
@@ -12,6 +12,7 @@ If needed, you can also manually update this file (provided the general structur
 ### [Rebranding]
 - Update corner radius for Buttons, TextFields, Snackbars, Dialogs and Drowndown menus.
 - Add new logos, replace existing ones.
+- Update color taxonomy.
 
 ## [Unreleased]